Improved on payment irregularities

+++ app/src/main/java/com/villagebanking/Model.kt
@@ -11,7 +11,7 @@ class Model {
     var accountHolderBankInfo: String = ""
     var accountHolderPin: String = ""
     var accountHolderPinHint: String = ""
-    var accountHoldersShare: Double = 0.00
+    var accountHoldersShare: Int = 0
     var accountHoldersLoanApp: Double = 0.0
     var accountHoldersCharges: Double = 0.0
     var accountHoldersApproved: String = ""
@@ -21,7 +21,7 @@ class Model {
     var transactionID: Int = 0
     var transactionName: String = ""
     var transactionMonth: String = ""
-    var transactionShares: Double = 0.00
+    var transactionShares: Int = 0
     var transactionShareAmount : Double = 0.00
     var transactionSharePayment: Double = 0.00
     var transactionShareDate: String = ""
